removed indexJs in components and renamed  independent styles in components for better understanding
