--- src/components/containers/section/index.js
@@ -1,1 +1,0 @@
-export { default } from './section'
