--- src/components/form/restoreForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './restoreForm'
