--- src/components/form/signInForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './signInForm'
