--- src/components/form/signUpForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './signUpForm'
