--- src/components/header/logo/index.js
@@ -1,1 +1,0 @@
-export { default } from './logo'
