--- src/components/header/userAvatar/index.js
@@ -1,1 +1,0 @@
-export { default } from './userAvatar'
