--- src/components/header/userName/index.js
@@ -1,1 +1,0 @@
-export { default } from './userName'
