+++ src/components/header/userName/userName.js
@@ -1,34 +1,4 @@
-import styled from 'styled-components'
-
-const UserWrapper = styled.div`
-  display: none;
-
-  @media screen and (min-width: 768px) {
-    display: flex;
-    flex-direction: column;
-    margin-left: auto;
-    margin-right: 16px;
-  }
-`
-
-const Text = styled.p`
-  @media screen and (min-width: 768px) {
-    font-size: 15px;
-    font-weight: 600;
-    line-height: 1.3;
-    color: var(--title-color);
-  }
-`
-
-const SubText = styled.span`
-  @media screen and (min-width: 768px) {
-    font-size: 13px;
-    font-weight: 500;
-    line-height: 1.3;
-    text-align: right;
-    color: var(--text-color);
-  }
-`
+import { UserWrapper, Text, SubText } from './userName.styles'
 
 const UserName = ({ user }) => {
   return (
