--- src/components/mainDoctor/emptyState/index.js
@@ -1,1 +1,0 @@
-export { default } from './emptyState'
