--- src/components/mainDoctor/index.js
@@ -1,1 +1,0 @@
-export { default } from './mainDoctor'
