--- src/components/mainDoctor/patientsList/index.js
@@ -1,1 +1,0 @@
-export { default } from './patientsList'
