--- src/components/mainPatient/appointmentsList/index.js
@@ -1,1 +1,0 @@
-export { default } from './appointmentsList'
