--- src/components/mainPatient/index.js
@@ -1,1 +1,0 @@
-export { default } from './mainPatient'
