--- src/components/preloader/index.js
@@ -1,1 +1,0 @@
-export { default } from './preloader'
