--- src/view/confirmView/index.js
@@ -1,1 +1,0 @@
-export { default } from './confirmView'
