+++ src/view/doctorView/doctorView.js
@@ -1,9 +1,9 @@
-import Header from '../../components/header'
-import Section from '../../components/containers/section'
-import Logo from '../../components/header/logo'
-import UserName from '../../components/header/userName'
-import UserAvatar from '../../components/header/userAvatar'
-import MainDoctor from '../../components/mainDoctor'
+import Header from '../../components/header/header'
+import Section from '../../components/containers/section/section'
+import Logo from '../../components/header/logo/logo'
+import UserName from '../../components/header/userName/userName'
+import UserAvatar from '../../components/header/userAvatar/userAvatar'
+import MainDoctor from '../../components/mainDoctor/mainDoctor'
 
 import doctorImg from '../../img/doctor-page/avatar.png'
