+++ src/view/patientView/patientView.js
@@ -1,9 +1,9 @@
-import Header from '../../components/header'
-import Section from '../../components/containers/section'
-import Logo from '../../components/header/logo'
-import UserName from '../../components/header/userName'
-import UserAvatar from '../../components/header/userAvatar'
-import MainPatient from '../../components/mainPatient'
+import Header from '../../components/header/header'
+import Section from '../../components/containers/section/section'
+import Logo from '../../components/header/logo/logo'
+import UserName from '../../components/header/userName/userName'
+import UserAvatar from '../../components/header/userAvatar/userAvatar'
+import MainPatient from '../../components/mainPatient/mainPatient'
 
 import patientImg from '../../img/doctor-page/user.png'
